Hackathons/ING-Global-Hackathon-2026/aml-validator-engine/requirements.txt
psycopg2-binary
confluent-kafka
pandas